feat: Add Suggestion model and update Suggestions model
This commit adds a new `Suggestion` model to represent a suggestion of a new paper topic from the related work. It includes fields for the topic and details of the suggestion. The `Suggestions` model is also updated to use a list of `Suggestion` objects instead of a generic list. This improves the clarity and structure of the code for managing suggestions.